refactor: rename InlineMarketCharts.tsx to InlineArtifactCards.tsx
The file holds all inline artifact card components (stock prices,
company overview, sector performance, web search, etc.) plus shared
constants and helpers. The old name was a holdover from when it only
had market data charts.
