+++ web/src/pages/ChatAgent/components/ActivityBlock.tsx
@@ -16,7 +16,7 @@ import {
   InlineSecFilingCard,
   InlineStockScreenerCard,
   InlineWebSearchCard,
-} from './charts/InlineMarketCharts';
+} from './charts/InlineArtifactCards';
 import { InlineAutomationCard } from './charts/InlineAutomationCards';
 import { InlinePreviewCard } from './charts/InlinePreviewCard';
 import { useTranslation } from 'react-i18next';
