+++ web/src/pages/ChatAgent/components/charts/InlineArtifactCards.tsx
